Move tests folder to the root folder
