+++ composer.json
@@ -11,17 +11,18 @@
   "minimum-stability": "dev",
   "autoload": {
     "psr-4": {
-      "WebAppId\\Content\\": "src"
+      "WebAppId\\Content\\": "src/"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "WebAppId\\Tests\\": "src/tests"
+      "WebAppId\\Content\\Tests\\": "tests/",
+      "WebAppId\\User\\Tests\\": "vendor/webappid/laravel-user/tests"
     }
   },
   "require": {
     "php": ">=7.2.0",
-    "illuminate/support": "6.*|7.*",
+    "illuminate/support": "^6.0|^7.0",
     "doctrine/dbal": "^2.5",
     "gumlet/php-image-resize": "^1.9",
     "nesbot/carbon": "^2.0",
@@ -29,10 +30,14 @@
     "webappid/laravel-user": "^3.0|^4.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^8.0",
-    "squizlabs/php_codesniffer": "^3.0",
+    "phpunit/phpunit": "^9.0",
     "mockery/mockery": "^1.0",
     "orchestra/testbench": "^4.0|^5.0"
+  },
+  "config": {
+    "preferred-install": "dist",
+    "sort-packages": true,
+    "optimize-autoloader": true
   },
   "extra": {
     "branch-alias": {
