+++ phpunit.xml
@@ -4,7 +4,7 @@
         bootstrap="vendor/autoload.php">
     <testsuites>
         <testsuite name="WebAppId Content Test Suite">
-            <directory>src/tests</directory>
+            <directory>tests</directory>
         </testsuite>
     </testsuites>
 </phpunit>
