+++ tests/Unit/Repositories/ContentStatusRepositoryTest.php
@@ -3,29 +3,29 @@
  * Created by LazyCrud - @DyanGalih <dyan.galih@gmail.com>
  */
 
-namespace WebAppId\Tests\Unit\Repositories;
+namespace WebAppId\Content\Tests\Unit\Repositories;
 
 use Illuminate\Contracts\Container\BindingResolutionException;
-use WebAppId\Content\Models\Tag;
-use WebAppId\Content\Repositories\Requests\TagRepositoryRequest;
-use WebAppId\Content\Repositories\TagRepository;
-use WebAppId\Tests\TestCase;
+use WebAppId\Content\Models\ContentStatus;
+use WebAppId\Content\Repositories\ContentStatusRepository;
+use WebAppId\Content\Repositories\Requests\ContentStatusRepositoryRequest;
+use WebAppId\Content\Tests\TestCase;
 use WebAppId\User\Tests\Unit\Repositories\UserRepositoryTest;
 
 /**
  * @author: Dyan Galih<dyan.galih@gmail.com>
- * Date: 05:01:58
+ * Date: 04:47:50
  * Time: 2020/04/22
- * Class TagServiceResponseList
- * @package WebAppId\Tests\Unit\Repositories
+ * Class ContentStatusServiceResponseList
+ * @package Tests\Unit\Repositories
  */
-class TagRepositoryTest extends TestCase
+class ContentStatusRepositoryTest extends TestCase
 {
 
     /**
-     * @var TagRepository
+     * @var ContentStatusRepository
      */
-    private $tagRepository;
+    private $contentStatusRepository;
 
     /**
      * @var UserRepositoryTest
@@ -36,20 +36,20 @@ class TagRepositoryTest extends TestCase
     {
         parent::__construct($name, $data, $dataName);
         try {
-            $this->tagRepository = $this->container->make(TagRepository::class);
+            $this->contentStatusRepository = $this->container->make(ContentStatusRepository::class);
             $this->userRepositoryTest = $this->container->make(UserRepositoryTest::class);
         } catch (BindingResolutionException $e) {
             report($e);
         }
     }
 
-    public function getDummy(int $no = 0): ?TagRepositoryRequest
+    public function getDummy(int $no = 0): ?ContentStatusRepositoryRequest
     {
         $dummy = null;
         try {
-            $dummy = $this->container->make(TagRepositoryRequest::class);
+            $dummy = $this->container->make(ContentStatusRepositoryRequest::class);
             $user = $this->container->call([$this->userRepositoryTest, 'testStore']);
-            $dummy->name = $this->getFaker()->text(255);
+            $dummy->name = $this->getFaker()->text(20);
             $dummy->user_id = $user->id;
 
         } catch (BindingResolutionException $e) {
@@ -58,32 +58,25 @@ class TagRepositoryTest extends TestCase
         return $dummy;
     }
 
-    public function testStore(int $no = 0): ?Tag
+    public function testStore(int $no = 0): ?ContentStatus
     {
-        $tagRepositoryRequest = $this->getDummy($no);
-        $result = $this->container->call([$this->tagRepository, 'store'], ['tagRepositoryRequest' => $tagRepositoryRequest]);
+        $contentStatusRepositoryRequest = $this->getDummy($no);
+        $result = $this->container->call([$this->contentStatusRepository, 'store'], ['contentStatusRepositoryRequest' => $contentStatusRepositoryRequest]);
         self::assertNotEquals(null, $result);
         return $result;
     }
 
     public function testGetById()
     {
-        $tag = $this->testStore();
-        $result = $this->container->call([$this->tagRepository, 'getById'], ['id' => $tag->id]);
-        self::assertNotEquals(null, $result);
-    }
-
-    public function testGetByName()
-    {
-        $tag = $this->testStore();
-        $result = $this->container->call([$this->tagRepository, 'getByName'], ['name' => $tag->name]);
+        $contentStatus = $this->testStore();
+        $result = $this->container->call([$this->contentStatusRepository, 'getById'], ['id' => $contentStatus->id]);
         self::assertNotEquals(null, $result);
     }
 
     public function testDelete()
     {
-        $tag = $this->testStore();
-        $result = $this->container->call([$this->tagRepository, 'delete'], ['id' => $tag->id]);
+        $contentStatus = $this->testStore();
+        $result = $this->container->call([$this->contentStatusRepository, 'delete'], ['id' => $contentStatus->id]);
         self::assertTrue($result);
     }
 
@@ -93,7 +86,7 @@ class TagRepositoryTest extends TestCase
             $this->testStore($i);
         }
 
-        $resultList = $this->container->call([$this->tagRepository, 'get']);
+        $resultList = $this->container->call([$this->contentStatusRepository, 'get']);
         self::assertGreaterThanOrEqual(1, count($resultList));
     }
 
@@ -103,37 +96,37 @@ class TagRepositoryTest extends TestCase
             $this->testStore($i);
         }
 
-        $result = $this->container->call([$this->tagRepository, 'getCount']);
+        $result = $this->container->call([$this->contentStatusRepository, 'getCount']);
         self::assertGreaterThanOrEqual(1, $result);
     }
 
     public function testUpdate()
     {
-        $tag = $this->testStore();
-        $tagRepositoryRequest = $this->getDummy(1);
-        $result = $this->container->call([$this->tagRepository, 'update'], ['id' => $tag->id, 'tagRepositoryRequest' => $tagRepositoryRequest]);
+        $contentStatus = $this->testStore();
+        $contentStatusRepositoryRequest = $this->getDummy(1);
+        $result = $this->container->call([$this->contentStatusRepository, 'update'], ['id' => $contentStatus->id, 'contentStatusRepositoryRequest' => $contentStatusRepositoryRequest]);
         self::assertNotEquals(null, $result);
     }
 
     public function testGetWhere()
     {
-        for ($i = 0; $i < $this->getFaker()->numberBetween(50, $this->getFaker()->numberBetween(50, 100)); $i++) {
+        for ($i = 0; $i < $this->getFaker()->numberBetween(10, $this->getFaker()->numberBetween(10, 20)); $i++) {
             $this->testStore($i);
         }
         $string = 'aiueo';
         $q = $string[$this->getFaker()->numberBetween(0, strlen($string) - 1)];
-        $result = $this->container->call([$this->tagRepository, 'get'], ['q' => $q]);
+        $result = $this->container->call([$this->contentStatusRepository, 'get'], ['q' => $q]);
         self::assertGreaterThanOrEqual(1, count($result));
     }
 
     public function testGetWhereCount()
     {
-        for ($i = 0; $i < $this->getFaker()->numberBetween(50, $this->getFaker()->numberBetween(50, 100)); $i++) {
+        for ($i = 0; $i < $this->getFaker()->numberBetween(10, $this->getFaker()->numberBetween(10, 30)); $i++) {
             $this->testStore($i);
         }
         $string = 'aiueo';
         $q = $string[$this->getFaker()->numberBetween(0, strlen($string) - 1)];
-        $result = $this->container->call([$this->tagRepository, 'getCount'], ['q' => $q]);
+        $result = $this->container->call([$this->contentStatusRepository, 'getCount'], ['q' => $q]);
         self::assertGreaterThanOrEqual(1, $result);
     }
 }
