+++ tests/Unit/Repositories/MimeTypeRepositoryTest.php
@@ -3,66 +3,53 @@
  * Created by LazyCrud - @DyanGalih <dyan.galih@gmail.com>
  */
 
-namespace WebAppId\Tests\Unit\Repositories;
+namespace WebAppId\Content\Tests\Unit\Repositories;
 
 use Illuminate\Contracts\Container\BindingResolutionException;
-use WebAppId\Content\Models\File;
-use WebAppId\Content\Repositories\FileRepository;
-use WebAppId\Content\Repositories\Requests\FileRepositoryRequest;
-use WebAppId\Tests\TestCase;
+use WebAppId\Content\Models\MimeType;
+use WebAppId\Content\Repositories\MimeTypeRepository;
+use WebAppId\Content\Repositories\Requests\MimeTypeRepositoryRequest;
+use WebAppId\Content\Tests\TestCase;
 use WebAppId\User\Tests\Unit\Repositories\UserRepositoryTest;
 
 /**
  * @author: Dyan Galih<dyan.galih@gmail.com>
- * Date: 04:10:49
+ * Date: 04:25:12
  * Time: 2020/04/22
- * Class FileServiceResponseList
- * @package Tests\Unit\Repositories
+ * Class MimeTypeServiceResponseList
+ * @package WebAppId\Content\Tests\Unit\Repositories
  */
-class FileRepositoryTest extends TestCase
+class MimeTypeRepositoryTest extends TestCase
 {
 
     /**
-     * @var FileRepository
+     * @var MimeTypeRepository
      */
-    private $fileRepository;
+    private $mimeTypeRepository;
 
     /**
      * @var UserRepositoryTest
      */
     private $userRepositoryTest;
 
-    /**
-     * @var MimeTypeRepositoryTest
-     */
-    private $mimeTypeRepositoryTest;
-
     public function __construct($name = null, array $data = [], $dataName = '')
     {
         parent::__construct($name, $data, $dataName);
         try {
-            $this->fileRepository = $this->container->make(FileRepository::class);
+            $this->mimeTypeRepository = $this->container->make(MimeTypeRepository::class);
             $this->userRepositoryTest = $this->container->make(UserRepositoryTest::class);
-            $this->mimeTypeRepositoryTest = $this->container->make(MimeTypeRepositoryTest::class);
         } catch (BindingResolutionException $e) {
             report($e);
         }
     }
 
-    public function getDummy(int $no = 0): ?FileRepositoryRequest
+    public function getDummy(int $no = 0): ?MimeTypeRepositoryRequest
     {
         $dummy = null;
         try {
-            $dummy = $this->container->make(FileRepositoryRequest::class);
+            $dummy = $this->container->make(MimeTypeRepositoryRequest::class);
             $user = $this->container->call([$this->userRepositoryTest, 'testStore']);
-            $mimeType = $this->container->call([$this->mimeTypeRepositoryTest, 'testStore']);
             $dummy->name = $this->getFaker()->text(255);
-            $dummy->description = $this->getFaker()->text(255);
-            $dummy->alt = $this->getFaker()->text(100);
-            $dummy->path = $this->getFaker()->text(255);
-            $dummy->mime_type_id = $mimeType->id;
-            $dummy->creator_id = $user->id;
-            $dummy->owner_id = $user->id;
             $dummy->user_id = $user->id;
 
         } catch (BindingResolutionException $e) {
@@ -71,32 +58,32 @@ class FileRepositoryTest extends TestCase
         return $dummy;
     }
 
-    public function testStore(int $no = 0): ?File
+    public function testStore(int $no = 0): ?MimeType
     {
-        $fileRepositoryRequest = $this->getDummy($no);
-        $result = $this->container->call([$this->fileRepository, 'store'], ['fileRepositoryRequest' => $fileRepositoryRequest]);
+        $mimeTypeRepositoryRequest = $this->getDummy($no);
+        $result = $this->container->call([$this->mimeTypeRepository, 'store'], ['mimeTypeRepositoryRequest' => $mimeTypeRepositoryRequest]);
         self::assertNotEquals(null, $result);
         return $result;
     }
 
     public function testGetById()
     {
-        $file = $this->testStore();
-        $result = $this->container->call([$this->fileRepository, 'getById'], ['id' => $file->id]);
+        $mimeType = $this->testStore();
+        $result = $this->container->call([$this->mimeTypeRepository, 'getById'], ['id' => $mimeType->id]);
         self::assertNotEquals(null, $result);
     }
 
     public function testGetByName()
     {
-        $file = $this->testStore();
-        $result = $this->container->call([$this->fileRepository, 'getByName'], ['name' => $file->name]);
+        $mimeType = $this->testStore();
+        $result = $this->container->call([$this->mimeTypeRepository, 'getByName'], ['name' => $mimeType->name]);
         self::assertNotEquals(null, $result);
     }
 
     public function testDelete()
     {
-        $file = $this->testStore();
-        $result = $this->container->call([$this->fileRepository, 'delete'], ['id' => $file->id]);
+        $mimeType = $this->testStore();
+        $result = $this->container->call([$this->mimeTypeRepository, 'delete'], ['id' => $mimeType->id]);
         self::assertTrue($result);
     }
 
@@ -106,7 +93,7 @@ class FileRepositoryTest extends TestCase
             $this->testStore($i);
         }
 
-        $resultList = $this->container->call([$this->fileRepository, 'get']);
+        $resultList = $this->container->call([$this->mimeTypeRepository, 'get']);
         self::assertGreaterThanOrEqual(1, count($resultList));
     }
 
@@ -116,15 +103,15 @@ class FileRepositoryTest extends TestCase
             $this->testStore($i);
         }
 
-        $result = $this->container->call([$this->fileRepository, 'getCount']);
+        $result = $this->container->call([$this->mimeTypeRepository, 'getCount']);
         self::assertGreaterThanOrEqual(1, $result);
     }
 
     public function testUpdate()
     {
-        $file = $this->testStore();
-        $fileRepositoryRequest = $this->getDummy(1);
-        $result = $this->container->call([$this->fileRepository, 'update'], ['id' => $file->id, 'fileRepositoryRequest' => $fileRepositoryRequest]);
+        $mimeType = $this->testStore();
+        $mimeTypeRepositoryRequest = $this->getDummy(1);
+        $result = $this->container->call([$this->mimeTypeRepository, 'update'], ['id' => $mimeType->id, 'mimeTypeRepositoryRequest' => $mimeTypeRepositoryRequest]);
         self::assertNotEquals(null, $result);
     }
 
@@ -135,7 +122,7 @@ class FileRepositoryTest extends TestCase
         }
         $string = 'aiueo';
         $q = $string[$this->getFaker()->numberBetween(0, strlen($string) - 1)];
-        $result = $this->container->call([$this->fileRepository, 'get'], ['q' => $q]);
+        $result = $this->container->call([$this->mimeTypeRepository, 'get'], ['q' => $q]);
         self::assertGreaterThanOrEqual(1, count($result));
     }
 
@@ -146,7 +133,7 @@ class FileRepositoryTest extends TestCase
         }
         $string = 'aiueo';
         $q = $string[$this->getFaker()->numberBetween(0, strlen($string) - 1)];
-        $result = $this->container->call([$this->fileRepository, 'getCount'], ['q' => $q]);
+        $result = $this->container->call([$this->mimeTypeRepository, 'getCount'], ['q' => $q]);
         self::assertGreaterThanOrEqual(1, $result);
     }
 }
