+++ tests/Unit/Repositories/TagRepositoryTest.php
@@ -3,29 +3,29 @@
  * Created by LazyCrud - @DyanGalih <dyan.galih@gmail.com>
  */
 
-namespace WebAppId\Tests\Unit\Repositories;
+namespace WebAppId\Content\Tests\Unit\Repositories;
 
 use Illuminate\Contracts\Container\BindingResolutionException;
-use WebAppId\Content\Models\ContentStatus;
-use WebAppId\Content\Repositories\ContentStatusRepository;
-use WebAppId\Content\Repositories\Requests\ContentStatusRepositoryRequest;
-use WebAppId\Tests\TestCase;
+use WebAppId\Content\Models\Tag;
+use WebAppId\Content\Repositories\Requests\TagRepositoryRequest;
+use WebAppId\Content\Repositories\TagRepository;
+use WebAppId\Content\Tests\TestCase;
 use WebAppId\User\Tests\Unit\Repositories\UserRepositoryTest;
 
 /**
  * @author: Dyan Galih<dyan.galih@gmail.com>
- * Date: 04:47:50
+ * Date: 05:01:58
  * Time: 2020/04/22
- * Class ContentStatusServiceResponseList
- * @package Tests\Unit\Repositories
+ * Class TagServiceResponseList
+ * @package WebAppId\Content\Tests\Unit\Repositories
  */
-class ContentStatusRepositoryTest extends TestCase
+class TagRepositoryTest extends TestCase
 {
 
     /**
-     * @var ContentStatusRepository
+     * @var TagRepository
      */
-    private $contentStatusRepository;
+    private $tagRepository;
 
     /**
      * @var UserRepositoryTest
@@ -36,20 +36,20 @@ class ContentStatusRepositoryTest extends TestCase
     {
         parent::__construct($name, $data, $dataName);
         try {
-            $this->contentStatusRepository = $this->container->make(ContentStatusRepository::class);
+            $this->tagRepository = $this->container->make(TagRepository::class);
             $this->userRepositoryTest = $this->container->make(UserRepositoryTest::class);
         } catch (BindingResolutionException $e) {
             report($e);
         }
     }
 
-    public function getDummy(int $no = 0): ?ContentStatusRepositoryRequest
+    public function getDummy(int $no = 0): ?TagRepositoryRequest
     {
         $dummy = null;
         try {
-            $dummy = $this->container->make(ContentStatusRepositoryRequest::class);
+            $dummy = $this->container->make(TagRepositoryRequest::class);
             $user = $this->container->call([$this->userRepositoryTest, 'testStore']);
-            $dummy->name = $this->getFaker()->text(20);
+            $dummy->name = $this->getFaker()->text(255);
             $dummy->user_id = $user->id;
 
         } catch (BindingResolutionException $e) {
@@ -58,25 +58,32 @@ class ContentStatusRepositoryTest extends TestCase
         return $dummy;
     }
 
-    public function testStore(int $no = 0): ?ContentStatus
+    public function testStore(int $no = 0): ?Tag
     {
-        $contentStatusRepositoryRequest = $this->getDummy($no);
-        $result = $this->container->call([$this->contentStatusRepository, 'store'], ['contentStatusRepositoryRequest' => $contentStatusRepositoryRequest]);
+        $tagRepositoryRequest = $this->getDummy($no);
+        $result = $this->container->call([$this->tagRepository, 'store'], ['tagRepositoryRequest' => $tagRepositoryRequest]);
         self::assertNotEquals(null, $result);
         return $result;
     }
 
     public function testGetById()
     {
-        $contentStatus = $this->testStore();
-        $result = $this->container->call([$this->contentStatusRepository, 'getById'], ['id' => $contentStatus->id]);
+        $tag = $this->testStore();
+        $result = $this->container->call([$this->tagRepository, 'getById'], ['id' => $tag->id]);
+        self::assertNotEquals(null, $result);
+    }
+
+    public function testGetByName()
+    {
+        $tag = $this->testStore();
+        $result = $this->container->call([$this->tagRepository, 'getByName'], ['name' => $tag->name]);
         self::assertNotEquals(null, $result);
     }
 
     public function testDelete()
     {
-        $contentStatus = $this->testStore();
-        $result = $this->container->call([$this->contentStatusRepository, 'delete'], ['id' => $contentStatus->id]);
+        $tag = $this->testStore();
+        $result = $this->container->call([$this->tagRepository, 'delete'], ['id' => $tag->id]);
         self::assertTrue($result);
     }
 
@@ -86,7 +93,7 @@ class ContentStatusRepositoryTest extends TestCase
             $this->testStore($i);
         }
 
-        $resultList = $this->container->call([$this->contentStatusRepository, 'get']);
+        $resultList = $this->container->call([$this->tagRepository, 'get']);
         self::assertGreaterThanOrEqual(1, count($resultList));
     }
 
@@ -96,37 +103,37 @@ class ContentStatusRepositoryTest extends TestCase
             $this->testStore($i);
         }
 
-        $result = $this->container->call([$this->contentStatusRepository, 'getCount']);
+        $result = $this->container->call([$this->tagRepository, 'getCount']);
         self::assertGreaterThanOrEqual(1, $result);
     }
 
     public function testUpdate()
     {
-        $contentStatus = $this->testStore();
-        $contentStatusRepositoryRequest = $this->getDummy(1);
-        $result = $this->container->call([$this->contentStatusRepository, 'update'], ['id' => $contentStatus->id, 'contentStatusRepositoryRequest' => $contentStatusRepositoryRequest]);
+        $tag = $this->testStore();
+        $tagRepositoryRequest = $this->getDummy(1);
+        $result = $this->container->call([$this->tagRepository, 'update'], ['id' => $tag->id, 'tagRepositoryRequest' => $tagRepositoryRequest]);
         self::assertNotEquals(null, $result);
     }
 
     public function testGetWhere()
     {
-        for ($i = 0; $i < $this->getFaker()->numberBetween(10, $this->getFaker()->numberBetween(10, 20)); $i++) {
+        for ($i = 0; $i < $this->getFaker()->numberBetween(50, $this->getFaker()->numberBetween(50, 100)); $i++) {
             $this->testStore($i);
         }
         $string = 'aiueo';
         $q = $string[$this->getFaker()->numberBetween(0, strlen($string) - 1)];
-        $result = $this->container->call([$this->contentStatusRepository, 'get'], ['q' => $q]);
+        $result = $this->container->call([$this->tagRepository, 'get'], ['q' => $q]);
         self::assertGreaterThanOrEqual(1, count($result));
     }
 
     public function testGetWhereCount()
     {
-        for ($i = 0; $i < $this->getFaker()->numberBetween(10, $this->getFaker()->numberBetween(10, 30)); $i++) {
+        for ($i = 0; $i < $this->getFaker()->numberBetween(50, $this->getFaker()->numberBetween(50, 100)); $i++) {
             $this->testStore($i);
         }
         $string = 'aiueo';
         $q = $string[$this->getFaker()->numberBetween(0, strlen($string) - 1)];
-        $result = $this->container->call([$this->contentStatusRepository, 'getCount'], ['q' => $q]);
+        $result = $this->container->call([$this->tagRepository, 'getCount'], ['q' => $q]);
         self::assertGreaterThanOrEqual(1, $result);
     }
 }
